Add Seperate C files for project
Here are all of the c files, I set them up a little bit to save us all time. Feel free to change it up if you want after you download them, but if you're changing anything major in the code you should do it in a separate branch before merging with the master branch so we can all know what's going on.  Also make sure to use the dos2Unix command if using MobaXTerm.